move 'defaultTo' tests to separate file
