+++ types/ramda/tsconfig.json
@@ -56,6 +56,7 @@
         "test/converge-tests.ts",
         "test/countBy-tests.ts",
         "test/curry-tests.ts",
-        "test/dec-tests.ts"
+        "test/dec-tests.ts",
+        "test/defaultTo-tests.ts"
     ]
 }
